+++ lib/route_localize.rb
@@ -6,7 +6,7 @@ module RouteLocalize
   module_function
 
   # Yields one or several route definitions if the route definition has a
-  # `localize` or `localize_url` scope
+  # `localize` or `localize_subdomain` scope
   #
   # The arguments it accepts are the arguments given to
   # `ActionDispatch::Routing::RouteSet`'s method `add_route`, with the addition
@@ -15,7 +15,7 @@ module RouteLocalize
   # The array it yields are the arguments accepted by `add_route` so that these
   # can be handed back to Rails to insert the yielded route.
   def translate_route(app, conditions, requirements, defaults, as, anchor, route_set)
-    locales = defaults.delete(:localize) || defaults.delete(:localize_url)
+    locales = defaults.delete(:localize) || defaults.delete(:localize_subdomain)
     if locales.present?
 
       # Makes sure the routes aren't created before i18n can read translations
@@ -44,4 +44,34 @@ module RouteLocalize
     end
   end
 
+
+  # Returns a translated path
+  # Example: "/trees/:id(.:format)" -> "/arbres/:id(.:format)", …
+  def translated_path(path, locale, by_subdomain: false)
+    path = path.dup
+
+    # Remove "(.:format)" in routes or "?args" if used elsewhere
+    final_options = path.slice!(/(\(.+\)|\?.*)$/)
+
+    segments = path.split('/').map do |segment|
+      translate_segment(segment, locale)
+    end
+
+    segments.unshift(":locale") unless by_subdomain
+    segments = segments.reject(&:blank?)
+
+    "/#{segments.join('/')}#{final_options}"
+  end
+
+  # Translates part of a path if it can
+  # Example: "trees" -> "arbres", ":id" -> ":id"
+  def translate_segment(segment, locale)
+    if segment =~ /^[a-z_0-9]+$/i
+      translation = I18n.t "routes.#{segment}", default: segment,
+                                                locale: locale
+      CGI.escape(translation)
+    else
+      segment
+    end
+  end
 end